requirements.txt
torch>=2.3
torchvision>=0.18
fastapi>=0.111
uvicorn[standard]>=0.30
python-multipart>=0.0.9
pydantic>=2.6
opencv-python>=4.9
numpy>=1.26
pillow>=10.2
scikit-learn>=1.5
matplotlib>=3.8
tqdm
